propchain_dex::dex

Struct PropertyDexRef

pub struct PropertyDexRef { /* private fields */ }

Implementations

impl PropertyDexRef

pub fn new( __ink_binding_0: String, __ink_binding_1: u128, __ink_binding_2: u128, __ink_binding_3: u32, ) -> CreateBuilder<<PropertyDex as ContractEnv>::Env, Self, Unset<<<PropertyDex as ContractEnv>::Env as Environment>::Hash>, Set<LimitParamsV2<<PropertyDex as ContractEnv>::Env>>, Unset<<<PropertyDex as ContractEnv>::Env as Environment>::Balance>, Set<ExecutionInput<ArgumentList<Argument<u32>, ArgumentList<Argument<u128>, ArgumentList<Argument<u128>, ArgumentList<Argument<String>, EmptyArgumentList>>>>>>, Unset<Salt>, Set<ReturnType<Self>>>

pub fn create_pool( &mut self, base_token: TokenId, quote_token: TokenId, fee_bips: u32, initial_base: u128, initial_quote: u128, ) -> Result<u64, Error>

pub fn try_create_pool( &mut self, base_token: TokenId, quote_token: TokenId, fee_bips: u32, initial_base: u128, initial_quote: u128, ) -> MessageResult<Result<u64, Error>>

pub fn add_liquidity( &mut self, pair_id: u64, amount_base: u128, amount_quote: u128, ) -> Result<u128, Error>

pub fn try_add_liquidity( &mut self, pair_id: u64, amount_base: u128, amount_quote: u128, ) -> MessageResult<Result<u128, Error>>

pub fn remove_liquidity( &mut self, pair_id: u64, shares: u128, ) -> Result<(u128, u128), Error>

pub fn try_remove_liquidity( &mut self, pair_id: u64, shares: u128, ) -> MessageResult<Result<(u128, u128), Error>>

pub fn swap_exact_base_for_quote( &mut self, pair_id: u64, amount_in: u128, min_quote_out: u128, ) -> Result<u128, Error>

pub fn try_swap_exact_base_for_quote( &mut self, pair_id: u64, amount_in: u128, min_quote_out: u128, ) -> MessageResult<Result<u128, Error>>

pub fn swap_exact_quote_for_base( &mut self, pair_id: u64, amount_in: u128, min_base_out: u128, ) -> Result<u128, Error>

pub fn try_swap_exact_quote_for_base( &mut self, pair_id: u64, amount_in: u128, min_base_out: u128, ) -> MessageResult<Result<u128, Error>>

pub fn place_order( &mut self, pair_id: u64, side: OrderSide, order_type: OrderType, time_in_force: TimeInForce, price: u128, amount: u128, trigger_price: Option<u128>, twap_interval: Option<u64>, reduce_only: bool, ) -> Result<u64, Error>

pub fn try_place_order( &mut self, pair_id: u64, side: OrderSide, order_type: OrderType, time_in_force: TimeInForce, price: u128, amount: u128, trigger_price: Option<u128>, twap_interval: Option<u64>, reduce_only: bool, ) -> MessageResult<Result<u64, Error>>

pub fn execute_order( &mut self, order_id: u64, requested_amount: u128, ) -> Result<u128, Error>

pub fn try_execute_order( &mut self, order_id: u64, requested_amount: u128, ) -> MessageResult<Result<u128, Error>>

pub fn match_orders( &mut self, maker_order_id: u64, taker_order_id: u64, amount: u128, ) -> Result<u128, Error>

pub fn try_match_orders( &mut self, maker_order_id: u64, taker_order_id: u64, amount: u128, ) -> MessageResult<Result<u128, Error>>

pub fn cancel_order(&mut self, order_id: u64) -> Result<(), Error>

pub fn try_cancel_order( &mut self, order_id: u64, ) -> MessageResult<Result<(), Error>>

pub fn configure_bridge_route( &mut self, destination_chain: ChainId, gas_estimate: u64, protocol_fee: u128, ) -> Result<(), Error>

pub fn try_configure_bridge_route( &mut self, destination_chain: ChainId, gas_estimate: u64, protocol_fee: u128, ) -> MessageResult<Result<(), Error>>

pub fn quote_cross_chain_trade( &self, destination_chain: ChainId, ) -> Result<BridgeFeeQuote, Error>

pub fn try_quote_cross_chain_trade( &self, destination_chain: ChainId, ) -> MessageResult<Result<BridgeFeeQuote, Error>>

pub fn create_cross_chain_trade( &mut self, pair_id: u64, order_id: Option<u64>, destination_chain: ChainId, recipient: <<PropertyDex as ContractEnv>::Env as Environment>::AccountId, amount_in: u128, min_amount_out: u128, ) -> Result<u64, Error>

pub fn try_create_cross_chain_trade( &mut self, pair_id: u64, order_id: Option<u64>, destination_chain: ChainId, recipient: <<PropertyDex as ContractEnv>::Env as Environment>::AccountId, amount_in: u128, min_amount_out: u128, ) -> MessageResult<Result<u64, Error>>

pub fn attach_bridge_request( &mut self, trade_id: u64, bridge_request_id: u64, ) -> Result<(), Error>

pub fn try_attach_bridge_request( &mut self, trade_id: u64, bridge_request_id: u64, ) -> MessageResult<Result<(), Error>>

pub fn finalize_cross_chain_trade(&mut self, trade_id: u64) -> Result<(), Error>

pub fn try_finalize_cross_chain_trade( &mut self, trade_id: u64, ) -> MessageResult<Result<(), Error>>

pub fn set_liquidity_mining_campaign( &mut self, emission_rate: u128, start_block: u64, end_block: u64, reward_token_symbol: String, ) -> Result<(), Error>

pub fn try_set_liquidity_mining_campaign( &mut self, emission_rate: u128, start_block: u64, end_block: u64, reward_token_symbol: String, ) -> MessageResult<Result<(), Error>>

pub fn create_trading_competition( &mut self, pair_id: Option<u64>, title: String, reward_amount: u128, start_block: u64, end_block: u64, min_trade_volume: u128, top_n: u32, reward_token_symbol: String, ) -> Result<u64, Error>

pub fn try_create_trading_competition( &mut self, pair_id: Option<u64>, title: String, reward_amount: u128, start_block: u64, end_block: u64, min_trade_volume: u128, top_n: u32, reward_token_symbol: String, ) -> MessageResult<Result<u64, Error>>

pub fn get_trading_competition( &self, competition_id: u64, ) -> Option<TradingCompetition>

pub fn try_get_trading_competition( &self, competition_id: u64, ) -> MessageResult<Option<TradingCompetition>>

pub fn get_competition_score( &self, competition_id: u64, account: <<PropertyDex as ContractEnv>::Env as Environment>::AccountId, ) -> u128

pub fn try_get_competition_score( &self, competition_id: u64, account: <<PropertyDex as ContractEnv>::Env as Environment>::AccountId, ) -> MessageResult<u128>

pub fn claim_competition_reward( &mut self, competition_id: u64, ) -> Result<u128, Error>

pub fn try_claim_competition_reward( &mut self, competition_id: u64, ) -> MessageResult<Result<u128, Error>>

pub fn list_competition_participants( &self, competition_id: u64, ) -> Vec<<<PropertyDex as ContractEnv>::Env as Environment>::AccountId>

pub fn try_list_competition_participants( &self, competition_id: u64, ) -> MessageResult<Vec<<<PropertyDex as ContractEnv>::Env as Environment>::AccountId>>

pub fn get_active_competitions(&self) -> Vec<TradingCompetition>

pub fn try_get_active_competitions( &self, ) -> MessageResult<Vec<TradingCompetition>>

pub fn get_competition_leaderboard( &self, competition_id: u64, ) -> Vec<(<<PropertyDex as ContractEnv>::Env as Environment>::AccountId, u128)>

pub fn try_get_competition_leaderboard( &self, competition_id: u64, ) -> MessageResult<Vec<(<<PropertyDex as ContractEnv>::Env as Environment>::AccountId, u128)>>

pub fn deactivate_competition( &mut self, competition_id: u64, ) -> Result<(), Error>

pub fn try_deactivate_competition( &mut self, competition_id: u64, ) -> MessageResult<Result<(), Error>>

pub fn activate_competition(&mut self, competition_id: u64) -> Result<(), Error>

pub fn try_activate_competition( &mut self, competition_id: u64, ) -> MessageResult<Result<(), Error>>

pub fn complete_competition(&mut self, competition_id: u64) -> Result<(), Error>

pub fn try_complete_competition( &mut self, competition_id: u64, ) -> MessageResult<Result<(), Error>>

pub fn get_competition_reward_token_symbol( &self, competition_id: u64, ) -> Option<String>

pub fn try_get_competition_reward_token_symbol( &self, competition_id: u64, ) -> MessageResult<Option<String>>

pub fn get_competition_reward_amount(&self, competition_id: u64) -> Option<u128>

pub fn try_get_competition_reward_amount( &self, competition_id: u64, ) -> MessageResult<Option<u128>>

pub fn get_competition_timer(&self, competition_id: u64) -> Option<(u64, u64)>

pub fn try_get_competition_timer( &self, competition_id: u64, ) -> MessageResult<Option<(u64, u64)>>

pub fn is_competition_active(&self, competition_id: u64) -> bool

pub fn try_is_competition_active( &self, competition_id: u64, ) -> MessageResult<bool>

pub fn set_competition_minimum_volume( &mut self, competition_id: u64, min_trade_volume: u128, ) -> Result<(), Error>

pub fn try_set_competition_minimum_volume( &mut self, competition_id: u64, min_trade_volume: u128, ) -> MessageResult<Result<(), Error>>

pub fn set_competition_reward_amount( &mut self, competition_id: u64, reward_amount: u128, ) -> Result<(), Error>

pub fn try_set_competition_reward_amount( &mut self, competition_id: u64, reward_amount: u128, ) -> MessageResult<Result<(), Error>>

pub fn get_competition_participant_score( &self, competition_id: u64, trader: <<PropertyDex as ContractEnv>::Env as Environment>::AccountId, ) -> u128

pub fn try_get_competition_participant_score( &self, competition_id: u64, trader: <<PropertyDex as ContractEnv>::Env as Environment>::AccountId, ) -> MessageResult<u128>

pub fn get_competition_participants_count(&self, competition_id: u64) -> u64

pub fn try_get_competition_participants_count( &self, competition_id: u64, ) -> MessageResult<u64>

pub fn get_competition_end_block(&self, competition_id: u64) -> Option<u64>

pub fn try_get_competition_end_block( &self, competition_id: u64, ) -> MessageResult<Option<u64>>

pub fn get_competition_start_block(&self, competition_id: u64) -> Option<u64>

pub fn try_get_competition_start_block( &self, competition_id: u64, ) -> MessageResult<Option<u64>>

pub fn get_competition_details( &self, competition_id: u64, ) -> Option<TradingCompetition>

pub fn try_get_competition_details( &self, competition_id: u64, ) -> MessageResult<Option<TradingCompetition>>

pub fn is_competition_reward_claimed( &self, competition_id: u64, trader: <<PropertyDex as ContractEnv>::Env as Environment>::AccountId, ) -> bool

pub fn try_is_competition_reward_claimed( &self, competition_id: u64, trader: <<PropertyDex as ContractEnv>::Env as Environment>::AccountId, ) -> MessageResult<bool>

pub fn get_competition_total_players(&self, competition_id: u64) -> u64

pub fn try_get_competition_total_players( &self, competition_id: u64, ) -> MessageResult<u64>

pub fn get_competition_top_n(&self, competition_id: u64) -> Option<u32>

pub fn try_get_competition_top_n( &self, competition_id: u64, ) -> MessageResult<Option<u32>>

pub fn get_competition_reward_symbol( &self, competition_id: u64, ) -> Option<String>

pub fn try_get_competition_reward_symbol( &self, competition_id: u64, ) -> MessageResult<Option<String>>

pub fn get_competition_title(&self, competition_id: u64) -> Option<String>

pub fn try_get_competition_title( &self, competition_id: u64, ) -> MessageResult<Option<String>>

pub fn get_competition_reward_info( &self, competition_id: u64, ) -> Option<(u128, String)>

pub fn try_get_competition_reward_info( &self, competition_id: u64, ) -> MessageResult<Option<(u128, String)>>

pub fn get_all_competitions(&self) -> Vec<TradingCompetition>

pub fn try_get_all_competitions(&self) -> MessageResult<Vec<TradingCompetition>>

pub fn tally_competition_leaderboard( &self, competition_id: u64, ) -> Vec<(<<PropertyDex as ContractEnv>::Env as Environment>::AccountId, u128)>

pub fn try_tally_competition_leaderboard( &self, competition_id: u64, ) -> MessageResult<Vec<(<<PropertyDex as ContractEnv>::Env as Environment>::AccountId, u128)>>

pub fn get_competition_status( &self, competition_id: u64, ) -> Option<(bool, u64, u64)>

pub fn try_get_competition_status( &self, competition_id: u64, ) -> MessageResult<Option<(bool, u64, u64)>>

pub fn get_competition_reward_state(&self, competition_id: u64) -> Option<bool>

pub fn try_get_competition_reward_state( &self, competition_id: u64, ) -> MessageResult<Option<bool>>

pub fn get_reward_share_for_trader( &self, competition_id: u64, trader: <<PropertyDex as ContractEnv>::Env as Environment>::AccountId, ) -> Option<u128>

pub fn try_get_reward_share_for_trader( &self, competition_id: u64, trader: <<PropertyDex as ContractEnv>::Env as Environment>::AccountId, ) -> MessageResult<Option<u128>>

pub fn get_competition_total_score(&self, competition_id: u64) -> u128

pub fn try_get_competition_total_score( &self, competition_id: u64, ) -> MessageResult<u128>

pub fn get_competition_count(&self) -> u64

pub fn try_get_competition_count(&self) -> MessageResult<u64>

pub fn get_competition_participant_reward_status( &self, competition_id: u64, trader: <<PropertyDex as ContractEnv>::Env as Environment>::AccountId, ) -> bool

pub fn try_get_competition_participant_reward_status( &self, competition_id: u64, trader: <<PropertyDex as ContractEnv>::Env as Environment>::AccountId, ) -> MessageResult<bool>

pub fn get_competition_reward_balance( &self, trader: <<PropertyDex as ContractEnv>::Env as Environment>::AccountId, ) -> u128

pub fn try_get_competition_reward_balance( &self, trader: <<PropertyDex as ContractEnv>::Env as Environment>::AccountId, ) -> MessageResult<u128>

pub fn get_competition_settings( &self, competition_id: u64, ) -> Option<(u32, u128)>

pub fn try_get_competition_settings( &self, competition_id: u64, ) -> MessageResult<Option<(u32, u128)>>

pub fn get_competition_details_by_title( &self, title: String, ) -> Vec<TradingCompetition>

pub fn try_get_competition_details_by_title( &self, title: String, ) -> MessageResult<Vec<TradingCompetition>>

pub fn get_competition_rewards_summary( &self, competition_id: u64, ) -> Option<(u128, bool)>

pub fn try_get_competition_rewards_summary( &self, competition_id: u64, ) -> MessageResult<Option<(u128, bool)>>

pub fn get_competition_status_summary( &self, competition_id: u64, ) -> Option<(bool, u64, u64, u128)>

pub fn try_get_competition_status_summary( &self, competition_id: u64, ) -> MessageResult<Option<(bool, u64, u64, u128)>>

pub fn get_competition_report( &self, competition_id: u64, ) -> Option<(String, u128, u64, u64)>

pub fn try_get_competition_report( &self, competition_id: u64, ) -> MessageResult<Option<(String, u128, u64, u64)>>

pub fn get_competition_metadata(&self, competition_id: u64) -> Option<String>

pub fn try_get_competition_metadata( &self, competition_id: u64, ) -> MessageResult<Option<String>>

pub fn get_competition_summary_for_user( &self, competition_id: u64, trader: <<PropertyDex as ContractEnv>::Env as Environment>::AccountId, ) -> Option<(u128, bool)>

pub fn try_get_competition_summary_for_user( &self, competition_id: u64, trader: <<PropertyDex as ContractEnv>::Env as Environment>::AccountId, ) -> MessageResult<Option<(u128, bool)>>

pub fn get_competition_summary_all(&self) -> Vec<(u64, bool, u128)>

pub fn try_get_competition_summary_all( &self, ) -> MessageResult<Vec<(u64, bool, u128)>>

pub fn get_competition_final_scores( &self, competition_id: u64, ) -> Vec<(<<PropertyDex as ContractEnv>::Env as Environment>::AccountId, u128)>

pub fn try_get_competition_final_scores( &self, competition_id: u64, ) -> MessageResult<Vec<(<<PropertyDex as ContractEnv>::Env as Environment>::AccountId, u128)>>

pub fn get_competition_trade_volume_goal( &self, competition_id: u64, ) -> Option<u128>

pub fn try_get_competition_trade_volume_goal( &self, competition_id: u64, ) -> MessageResult<Option<u128>>

pub fn get_competition_reward_distribution( &self, competition_id: u64, ) -> Option<(u128, u32)>

pub fn try_get_competition_reward_distribution( &self, competition_id: u64, ) -> MessageResult<Option<(u128, u32)>>

pub fn get_competition_details_for_dashboard( &self, competition_id: u64, ) -> Option<(String, bool, u128)>

pub fn try_get_competition_details_for_dashboard( &self, competition_id: u64, ) -> MessageResult<Option<(String, bool, u128)>>

pub fn get_competition_history(&self) -> Vec<TradingCompetition>

pub fn try_get_competition_history( &self, ) -> MessageResult<Vec<TradingCompetition>>

pub fn get_competition_description(&self, competition_id: u64) -> Option<String>

pub fn try_get_competition_description( &self, competition_id: u64, ) -> MessageResult<Option<String>>

pub fn get_competition_rank( &self, competition_id: u64, trader: <<PropertyDex as ContractEnv>::Env as Environment>::AccountId, ) -> Option<u64>

pub fn try_get_competition_rank( &self, competition_id: u64, trader: <<PropertyDex as ContractEnv>::Env as Environment>::AccountId, ) -> MessageResult<Option<u64>>

pub fn get_competition_top_scores( &self, competition_id: u64, ) -> Vec<(<<PropertyDex as ContractEnv>::Env as Environment>::AccountId, u128)>

pub fn try_get_competition_top_scores( &self, competition_id: u64, ) -> MessageResult<Vec<(<<PropertyDex as ContractEnv>::Env as Environment>::AccountId, u128)>>

pub fn get_competition_active_status(&self, competition_id: u64) -> bool

pub fn try_get_competition_active_status( &self, competition_id: u64, ) -> MessageResult<bool>

pub fn get_competition_admin( &self, _competition_id: u64, ) -> <<PropertyDex as ContractEnv>::Env as Environment>::AccountId

pub fn try_get_competition_admin( &self, _competition_id: u64, ) -> MessageResult<<<PropertyDex as ContractEnv>::Env as Environment>::AccountId>

pub fn claim_liquidity_rewards(&mut self, pair_id: u64) -> Result<u128, Error>

pub fn try_claim_liquidity_rewards( &mut self, pair_id: u64, ) -> MessageResult<Result<u128, Error>>

pub fn pending_liquidity_rewards( &self, pair_id: u64, provider: <<PropertyDex as ContractEnv>::Env as Environment>::AccountId, ) -> Result<u128, Error>

pub fn try_pending_liquidity_rewards( &self, pair_id: u64, provider: <<PropertyDex as ContractEnv>::Env as Environment>::AccountId, ) -> MessageResult<Result<u128, Error>>

pub fn get_liquidity_position( &self, pair_id: u64, provider: <<PropertyDex as ContractEnv>::Env as Environment>::AccountId, ) -> LiquidityPosition

pub fn try_get_liquidity_position( &self, pair_id: u64, provider: <<PropertyDex as ContractEnv>::Env as Environment>::AccountId, ) -> MessageResult<LiquidityPosition>

pub fn get_liquidity_mining_campaign(&self) -> LiquidityMiningCampaign

pub fn try_get_liquidity_mining_campaign( &self, ) -> MessageResult<LiquidityMiningCampaign>

pub fn create_governance_proposal( &mut self, title: String, description_hash: [u8; 32], new_fee_bips: Option<u32>, new_emission_rate: Option<u128>, duration_blocks: u64, ) -> Result<u64, Error>

pub fn try_create_governance_proposal( &mut self, title: String, description_hash: [u8; 32], new_fee_bips: Option<u32>, new_emission_rate: Option<u128>, duration_blocks: u64, ) -> MessageResult<Result<u64, Error>>

pub fn vote_on_proposal( &mut self, proposal_id: u64, support: bool, ) -> Result<(), Error>

pub fn try_vote_on_proposal( &mut self, proposal_id: u64, support: bool, ) -> MessageResult<Result<(), Error>>

pub fn execute_governance_proposal( &mut self, proposal_id: u64, ) -> Result<bool, Error>

pub fn try_execute_governance_proposal( &mut self, proposal_id: u64, ) -> MessageResult<Result<bool, Error>>

pub fn get_pool(&self, pair_id: u64) -> Option<LiquidityPool>

pub fn try_get_pool(&self, pair_id: u64) -> MessageResult<Option<LiquidityPool>>

pub fn get_order(&self, order_id: u64) -> Option<TradingOrder>

pub fn try_get_order( &self, order_id: u64, ) -> MessageResult<Option<TradingOrder>>

pub fn get_pair_analytics(&self, pair_id: u64) -> Option<PairAnalytics>

pub fn try_get_pair_analytics( &self, pair_id: u64, ) -> MessageResult<Option<PairAnalytics>>

pub fn discover_price(&self, pair_id: u64) -> Result<u128, Error>

pub fn try_discover_price( &self, pair_id: u64, ) -> MessageResult<Result<u128, Error>>

pub fn get_portfolio_snapshot( &self, account: <<PropertyDex as ContractEnv>::Env as Environment>::AccountId, ) -> PortfolioSnapshot

pub fn try_get_portfolio_snapshot( &self, account: <<PropertyDex as ContractEnv>::Env as Environment>::AccountId, ) -> MessageResult<PortfolioSnapshot>

pub fn calculate_price_impact( &self, pair_id: u64, side: OrderSide, amount_in: u128, ) -> Result<(u32, u128), Error>

Calculate the expected price impact for a given trade amount. Returns the price impact in basis points (bips) and the expected output amount. This allows users to check the impact before executing a trade.

pub fn try_calculate_price_impact( &self, pair_id: u64, side: OrderSide, amount_in: u128, ) -> MessageResult<Result<(u32, u128), Error>>

Calculate the expected price impact for a given trade amount. Returns the price impact in basis points (bips) and the expected output amount. This allows users to check the impact before executing a trade.

pub fn get_governance_balance( &self, account: <<PropertyDex as ContractEnv>::Env as Environment>::AccountId, ) -> u128

pub fn try_get_governance_balance( &self, account: <<PropertyDex as ContractEnv>::Env as Environment>::AccountId, ) -> MessageResult<u128>

pub fn get_admin_timelock_delay(&self) -> u64

pub fn try_get_admin_timelock_delay(&self) -> MessageResult<u64>

pub fn set_admin_timelock_delay( &mut self, delay_blocks: u64, ) -> Result<(), Error>

pub fn try_set_admin_timelock_delay( &mut self, delay_blocks: u64, ) -> MessageResult<Result<(), Error>>

pub fn schedule_bridge_route_update( &mut self, destination_chain: ChainId, gas_estimate: u64, protocol_fee: u128, ) -> Result<u64, Error>

pub fn try_schedule_bridge_route_update( &mut self, destination_chain: ChainId, gas_estimate: u64, protocol_fee: u128, ) -> MessageResult<Result<u64, Error>>

pub fn schedule_liquidity_mining_update( &mut self, emission_rate: u128, start_block: u64, end_block: u64, reward_token_symbol: String, ) -> Result<u64, Error>

pub fn try_schedule_liquidity_mining_update( &mut self, emission_rate: u128, start_block: u64, end_block: u64, reward_token_symbol: String, ) -> MessageResult<Result<u64, Error>>

pub fn schedule_timelock_delay_update( &mut self, delay_blocks: u64, ) -> Result<u64, Error>

pub fn try_schedule_timelock_delay_update( &mut self, delay_blocks: u64, ) -> MessageResult<Result<u64, Error>>

pub fn execute_admin_action(&mut self, action_id: u64) -> Result<(), Error>

pub fn try_execute_admin_action( &mut self, action_id: u64, ) -> MessageResult<Result<(), Error>>

pub fn cancel_admin_action(&mut self, action_id: u64) -> Result<(), Error>

pub fn try_cancel_admin_action( &mut self, action_id: u64, ) -> MessageResult<Result<(), Error>>

pub fn get_scheduled_admin_action( &self, action_id: u64, ) -> Option<PendingAdminAction>

pub fn try_get_scheduled_admin_action( &self, action_id: u64, ) -> MessageResult<Option<PendingAdminAction>>

pub fn get_order_book_snapshot( &self, pair_id: u64, max_levels: u32, ) -> Result<OrderBookSnapshot, Error>

pub fn try_get_order_book_snapshot( &self, pair_id: u64, max_levels: u32, ) -> MessageResult<Result<OrderBookSnapshot, Error>>

pub fn get_order_book_levels( &self, pair_id: u64, side: OrderSide, max_levels: u32, ) -> Result<Vec<OrderBookLevel>, Error>

pub fn try_get_order_book_levels( &self, pair_id: u64, side: OrderSide, max_levels: u32, ) -> MessageResult<Result<Vec<OrderBookLevel>, Error>>

pub fn get_trading_statistics(&self) -> TradingStatistics

Get comprehensive trading statistics across all pairs

pub fn try_get_trading_statistics(&self) -> MessageResult<TradingStatistics>

Get comprehensive trading statistics across all pairs

pub fn get_price_history(&self, pair_id: u64) -> Option<PriceHistory>

Get price history summary for a trading pair

pub fn try_get_price_history( &self, pair_id: u64, ) -> MessageResult<Option<PriceHistory>>

Get price history summary for a trading pair

pub fn get_volume_analytics(&self, pair_id: u64) -> Option<VolumeAnalytics>

Get volume analytics for a trading pair

pub fn try_get_volume_analytics( &self, pair_id: u64, ) -> MessageResult<Option<VolumeAnalytics>>

Get volume analytics for a trading pair

pub fn get_all_pair_analytics(&self) -> Vec<PairAnalytics>

Get analytics for all trading pairs

pub fn try_get_all_pair_analytics(&self) -> MessageResult<Vec<PairAnalytics>>

Get analytics for all trading pairs